refactor: split napi docs bindings
