+++ crates/ox_content_napi/src/docs_bindings/paths.rs
@@ -1,0 +1,5 @@
+use std::path::Path;
+
+pub(super) fn path_to_string(path: &Path) -> String {
+    path.to_string_lossy().to_string()
+}
